Extract native transcript view
