+++ docs/DECISIONS.md
@@ -194,3 +194,4 @@
 - Agent progress status copy lives in `WorkspaceAgentStatusBuilder`. `WorkspaceModel` owns when agent progress arrives and when the top bar refreshes, while the builder owns event-kind-to-status labels such as Queued, Running, Review, Streaming, Finishing, and Failed so progress copy stays directly tested and aligned with top-bar presentation semantics.
 - Slash-command local transcript copy lives in `WorkspaceSlashCommandTranscriptPlanner`. `WorkspaceModel` still owns command side effects such as renaming, scheduling, SSH project creation, and tool dispatch, but the user-visible success/failure records for `/help`, `/status`, `/mode`, `/model`, `/rename`, `/project rename`, `/ssh`, `/follow-up`, `/workspace-check`, invalid commands, and unknown commands are pure planner outputs with focused tests.
 - Per-turn agent runner configuration lives in `WorkspaceAgentRunContextBuilder`. `WorkspaceModel` selects the current project/browser/memory/MCP/Computer Use state and asks the builder for a configured `AgentRunner`; the builder owns local versus SSH Remote base tools, optional plan/browser/Computer Use/memory/MCP tool definitions, and override composition. Memory tool execution and saved-memory event detection live together in `WorkspaceMemoryRememberToolExecutor`.
+- Transcript timeline layout lives in `QuillCodeTranscriptView`, not the workspace shell. The workspace shell decides whether the transcript pane is visible and routes callbacks; the transcript view owns Find bar placement, empty state, context banner/runtime issue/review placement, message and tool-card timeline rows, active Find highlighting, scroll-to-match behavior, and tool-card copy fallback. `WorkspaceSwiftUIView.swift` should stay focused on chrome composition, modal presentation, and typed command routing.
